+++ src/server/include/nms_core.h
@@ -1041,8 +1041,8 @@ void NXCORE_EXPORTABLE QueueSQLRequest(const TCHAR *query, int bindCount, int *s
 void QueueIDataInsert(time_t timestamp, UINT32 nodeId, UINT32 dciId, const TCHAR *rawValue, const TCHAR *transformedValue);
 void QueueRawDciDataUpdate(time_t timestamp, UINT32 dciId, const TCHAR *rawValue, const TCHAR *transformedValue);
 void QueueRawDciDataDelete(UINT32 dciId);
-size_t GetIDataWriterQueueSize();
-int GetRawDataWriterQueueSize();
+INT64 GetIDataWriterQueueSize();
+INT64 GetRawDataWriterQueueSize();
 void StartDBWriter();
 void StopDBWriter();
